Allow get[io]speed() to return a baud rate; ignore BOTHER for set
set[io]speed() already falls back to assuming a baud rate if flags
other than the legacy values are set.  This makes sense as those
values really are quite unlikely to appear in real life (depending on
architecture, either 1-15, 4097-4111 baud or 1-30 baud), although it
is still a heuristic.

Either way, return the actual baud rate rather than BOTHER, which is
completely useless to userspace.

Similarly, if user space passes BOTHER to set[io]speed(), treat it as
a baud rate (4096 or 31). Either of those baud rates are marginally
more plausible than the ones above, the first one being a power of 2
and the second one being used in amateur radio.