Disable float128 support for x86_64, x86
This disabled float128 support for x86_64, x86. This is a simplied
revert of commit c86ed71d633c22d6f638576f7660c52a5f783d66

Author: Joseph Myers <joseph@codesourcery.com>
Date:   Mon Jun 26 22:01:27 2017 +0000

    Add float128 support for x86_64, x86.

The reason to revert it is we are using compiler-rt instead of libgcc.
Some builtins are not ready in compiler-rt, as a result, we hit
some problems like b:187784697

[Adrian: forward-ported to glibc 2.35]