Add dlopen_with_offset to glibc.
This patchset is used in Google internal and is needed for
Loonix, https://crbug.com/1015890.

The patch is a squashed diff of following commits:
https://sourceware.org/git/?p=glibc.git;a=commitdiff;h=0010a400dcb13304c9081b0f9ad156ce88178320
https://sourceware.org/git/?p=glibc.git;a=commit;h=024f0eb409268e704a83cae2c1511d31167c6f49
https://sourceware.org/git/?p=glibc.git;a=commit;h=295883290d7f26f483e574e80a851cea2d639d12
https://sourceware.org/git/?p=glibc.git;a=commit;h=771d3bf7065f6c25219cf206816b695a628a4618
https://sourceware.org/git/?p=glibc.git;a=commit;h=a6241b657547f0f708629bc254ebcaafb6f25074
https://sourceware.org/git/?p=glibc.git;a=commit;h=c0342bb26e0251dc2d7ca00a884ea831705e32b4

Then modified to support LARGEFILE 32 bit builds and to only call lseek
if the offset is larger than 0 (to avoid irritating sandbox policy).

This is requested by Loonix for use in Google internal.

[Adrian: forward ported to glibc 2.35]