math: Use an improved algorithm for hypot
This implementation is based on the 'An Improved Algorithm for
hypot(a,b)' by Carlos F. Borges [1] using the MyHypot3 with the
following changes:

 - Handle qNaN and sNaN
 - Tune the 'widely varying operands' to avoid spurious underflow
   due the multiplication and fix the return value for upwards
   rounding mode.
 - Handle required underflow exception for denormal results.

The main advantage of the new algorithm is its precision.  With a
random 1e9 input pairs in the range of [DBL_MIN, DBL_MAX], glibc
current implementation shows around 0.34% results with an error of
1 ulp (3424869 results) while the new implementation only shows
0.002% of total (18851).

The performance result are also only slight worse than current
implementation.  On x86_64 (i7-4790K) with gcc 9.2.1:

Before:

$ taskset -c 4 ./testrun.sh ./benchtests/bench-hypot
  "hypot": {
   "workload-random": {
    "duration": 4.04086e+09,
    "iterations": 1.48e+08,
    "reciprocal-throughput": 25.7887,
    "latency": 28.8176,
    "max-throughput": 3.87767e+07,
    "min-throughput": 3.4701e+07
   }
  }

After:

$ taskset -c 4 ./testrun.sh ./benchtests/bench-hypot
  "hypot": {
   "workload-random": {
    "duration": 4.03499e+09,
    "iterations": 1.34e+08,
    "reciprocal-throughput": 30.0049,
    "latency": 30.2188,
    "max-throughput": 3.33279e+07,
    "min-throughput": 3.30919e+07
   }
  }

Co-Authored-By: Paul Zimmermann <Paul.Zimmermann@inria.fr>

Checked on x86_64-linux-gnu and aarch64-linux-gnu.

[1] https://arxiv.org/pdf/1904.09481.pdf