socket: Add socket timestamp test
The test check if SO_TIMESTAMP and SO_TIMESTAMPNS is correctly generated
by recvmsg and recvmmsg for both _TIME_BITS.  The recvmsg()/recvmmsg()
is called with different ancilliary data buffers, to check if the
conversion for 64-bit time on 32-bit legacy ABIs are handled correctly.

For legacy ABI with 32-bit time_t support, depending of the remmaning
ancially data space the timestamp can not be appended and MSG_TRUNC
is returned instead.

Checked on x86_64-linux and on i686-linux-gnu on both 5.11 and
4.15 kernel.