clone-internal.c: make clone3 syscall optional
This patch was originally carried in Gentoo to make clone3
optional to avoid breakage for CEF/steam/zoom, and others.

Gentoo made clone3 mandatory but ChromeOS is not there yet,
so we carry this trivial patch to allow us to keep clone3
disabled for a while longer.

See b:288928916 which is tracking the clone3 migration.

Original commit message:

We're disabling clone3 for now _CONDITIONALLY_ (not by default) to allow
compatibility with applications using older Electron.

Use -DGENTOO_USE_CLONE3 to enable clone3 for now. In future, we will
revert back to always using clone3.

This was impacting e.g. Discord and Skype. This patch stops glibc from using
clone3 internally (which is the only real use of it) and falls back to the old
behaviour.

Specifically, we want https://github.com/electron/electron/pull/31091
to work its way downstream to various Electron applications.

https://bugs.gentoo.org/819045
https://bugs.gentoo.org/827386

https://git.launchpad.net/~ubuntu-core-dev/ubuntu/+source/glibc/plain/debian/patches/ubuntu/disable-clone3.patch

This is the same as the patch that was considered but ultimately rejected
for 2.34 because Docker got sorted out in time:
https://patchwork.ozlabs.org/project/glibc/patch/87eebkf8ph.fsf@oldenburg.str.redhat.com/.

Signed-off-by: Sam James <sam@gentoo.org>