Always define __USE_TIME_BITS64
It was raised on libc-help [1] that some Linux kernel interfaces expects
the libc to define __USE_TIME_BITS64 to indicate the time_t size for the
kABI.  Different than defined by the initial y2038 design document [2],
the __USE_TIME_BITS64 is only defined for ABIs that support more than
time_t size (by defining the _TIME_BITS for each module).

The 64 bit time_t redirects are now used for __USE_TIME_BITS64 along
with __TIMESIZE equal to 32 (since it is the size of the default ABI).
There is no expected change in semantic or code generation.

On the libc-help it was raised that this might break applications that
use __USE_TIME_BITS64 to implie redirection.  In this case, since the
code is already using an internal interface, it can also checks for
__TIMESIZE similar to what installed headers do.

Checked on x86_64-linux-gnu, i686-linux-gnu, aarch64-linux-gnu,
and arm-linux-gnueabi

[1] https://sourceware.org/pipermail/libc-help/2024-January/006557.html
[2] https://sourceware.org/glibc/wiki/Y2038ProofnessDesign