linux: Add openat2 (BZ 31664)
The openat2 syscall was added on Linux 5.6, as an extension of openat.
Different than other open-like functions, the kernel only provides the
LFS variant (so files larger than 4GB always succeed, as other functions
with offset larger than off_t).  Also, imilar to other open functions,
the new symbol is a cancellable entrypoint.

The test case added only stresses some of the syscalls provided
functionalities and it is based on an existent kernel self-test.  Since
the prototype does not use variadic arguments, there is no need to add a
fortify wrapper to catch wrong usages.

The documentation mostly defers to the Linux manual pages.

Checked on x86_64-linux-gnu.