linux: Add support for x86_64 getrandom vDSO
The vDSO symbol operates in an opaque state, where it should be
allocated with mmap using the kernel-provided protection and flags.
It is done by issuing the getrandom vDSO symbol with the additional
vgetrandom_opaque_params argument.  Multiple states are allocated, so
getrandom provides per-thread support without locking (the current
strategy is to initially allocate the number of CPUs).

To keep track and re-use the opaque state when threads exit, getrandom
keeps track of the available ones in a FIFO.  When there is no available
opaque state, a new one is allocated.

Also, to make the vDSO call async-signal-safe, a strategy similar to
the vDSO is used where the allocated buffer is marked as taken by
setting LSB to 1 and cleared after the vDSO call.

Co-authored-by: Jason A. Donenfeld <Jason@zx2c4.com>