linux: Add support for getrandom vDSO
Linux 6.11 will support calling getrandom() from the vDSO (commit
4ad10a5f5f78a5b3e525a63bd075a4eb1139dde1). It operates on a thread-local
opaque state allocated with mmap using flags specified by the vDSO.

Multiple states are allocated at once, as many as fit into a page, and
these are held in an array of available states to be doled out to each
thread upon first use, and recycled when a thread terminates.  As these
states run low, more are allocated.

To make this procedure async-signal-safe, a simple guard is used in the
LSB of the opaque state address, falling back to the syscall if there's
reentrancy contention.

Also, _Fork() is handled by blocking signals on opaque state allocation
(so _Fork() always see a consistent state even if interrupts a
getrandom() call) and by iterating over the thread stack cache on
reclaim_stack.  Each opaque states will be either in the free states
list (grnd_alloc.states) or allocated to a running thread.

It's currently enabled for x86_64. As the kernel support gains more
platforms (arm64 is in the works), this can be easily turned on for
those.

Checked on x86_64-linux-gnu.

Co-authored-by: Jason A. Donenfeld <Jason@zx2c4.com>

NB: This potentially makes getrandom a non-cancellable entrypoint
    if the vDSO syscall is called for any reason. This is something
    we're going to address later, after this lands.