linux: Add support for getrandom vDSO
Linux 6.11 will support calling getrandom() from the vDSO (commit
4ad10a5f5f78a5b3e525a63bd075a4eb1139dde1). It operates on a thread-local
opaque state allocated with mmap using flags specified by the vDSO.

Multiple states are allocated at once, as many as fit into a page, and
these are held in an array of available states to be doled out to each
thread upon first use, and recycled when a thread terminates.  As these
states run low, more are allocated.

To make this procedure async-signal-safe, a simple guard is used in the
LSB of the opaque state address, falling back to the syscall if there's
reentrancy contention.

Also, _Fork() is handled by blocking signals on opaque state allocation
(so _Fork() always sees a consistent state even if it interrupts a
getrandom() call) and by iterating over the thread stack cache on
reclaim_stack.  Each opaque state will be in the free states list
(grnd_alloc.states) or allocated to a running thread.

The cancellation is handled by always using GRND_NONBLOCK flags while
calling the vDSO, and falling back to the cancellable syscall if the
kernel returns EAGAIN (would block).   Since getrandom is not defined
by POSIX and cancellation is supported as an extension, the cancellation
is handled as 'may occur' instead of 'shall occur' [1], meaning that
if vDSO does not block (the expected behavior) getrandom will not
act as a cancellation entrypoint.  It avoids a pthread_testcancel call
on the fast path (different than 'shall occur' functions, like sem_wait()).

It is currently enabled for x86_64. And the kernel's random.git tree has
support pending for arm64, powerpc32, powerpc64, and loongarch64.

Checked on x86_64-linux-gnu.

Co-authored-by: Jason A. Donenfeld <Jason@zx2c4.com>

[1] https://pubs.opengroup.org/onlinepubs/9799919799/nframe.html