Do not use -Wp to disable fortify (BZ 31928)
The -Wp does not work correclty if the compiler is configured to enable
fortify by default, since it bypass the compiler driver (which defines
the fortify flags in this case).

I checked with a build for x86_64-linux-gnu, i686-linux-gnu,
aarch64-linux-gnu, s390x-linux-gnu, and riscv64-linux-gnu with
gcc-13 that enables the fortify by default.