math: Add internal roundeven_finite
Some CORE-MATH routines uses roundeven and most of ISA do not have
an specific instruction for the operation.  In this case, the call
will be routed to generic implementation.

However, if the ISA does support round() and ctz() there is a better
alternative (as used by CORE-MATH).

This patch adds such optimization and also enables it on powerpc.
On a power10 it shows the following improvement:

expm1f
latency       master       patched       improvement
power10       4.3910        2.6595            39.43%
power10       15.8168      11.9993            24.14%

Checked on powerpc64le-linux-gnu and aarch64-linux-gnu.