math: Add inf support on gen-auto-libm-tests.c
For some correctly rounded inputs where infinity might generate
a number (like atanf), comparing to a pre-defined constant does not
yield the expected result in all roundind modes.

The most straightforward way to handle it would be to get the expected
result from mpfr, where it handles all the rounding mode.

The patch also adds support for infinity results, that will be used
to check for some complex implementations.