linux: Add mseal syscall support
It has been added on Linux 6.10 (8be7258aad44b5e25977a98db136f677fa6f4370)
as a way to block operations such as mapping, moving to another location,
shrinking the size, expanding the size, or modifying it to a pre-existent
memory mapping.

Although the system only works on 64-bit CPU, the entrypoint was added
for all ABIs (since kernel might eventually implement it to additional
ones and/or the abi can execute on a 64-bit kernel).

Checked on x86_64-linux-gnu.