aarch64: Add glibc.cpu.aarch64_gcs tunable
This tunable is for controlling the GCS status. It is used to work
out the argument to the PR_SET_SHADOW_STACK_STATUS prctl syscall.

By default it is 0, so GCS is disabled, value 1 will enable GCS.

The status is stored into GL(dl_aarch64_gcs) early and only applied
later, since enabling GCS is tricky: it must happen on a top level
stack frame. Using GL instead of GLRO because it may need updates
depending on loaded libraries that happen after readonly protection
is applied, however library marking based GCS setting is not yet
implemented.

Describe new tunable in the manual.