linux: Add support for PT_GNU_MUTABLE
Since memory sealing is applied for all the PT_LOAD segments,
a new program header is also supported so users can define a
region where sealing should not be applied (so users can later
initialize and change memory protection). The special section,
PT_GNU_MUTABLE (reference implemented at [1]),  marks a memory
region that should not be sealed if the GNU_PROPERTY_MEMORY_SEAL
attribute is present.

The section name starts with ".gnu.mutable" and has an alignment
and size of the defined maximum page size (-z max-page-size linker
option). For instance the code:

  #define GNU_MUTABLE_SECTION_NAME       ".gnu.mutable"

  unsigned char mutable_array1[64]
    __attribute__ ((section (GNU_MUTABLE_SECTION_NAME)))
       = { 0 };
  unsigned char mutable_array2[32]
    __attribute__ ((section (GNU_MUTABLE_SECTION_NAME)))
       = { 0 };

places both 'mutable_array1' and 'mutable_array2' on a page
aligned memory region with the size of a page.

The linker sets the alignment and size to simplify support for ABIs
with multiple page sizes, otherwise user would need to know the
maximum page size to correctly define the alignment and size of the
variable.

Checked on aarch64-linux-gnu and x86_64-linux-gnu.

[1] https://sourceware.org/git/?p=binutils-gdb.git;a=shortlog;h=refs/heads/azanella/pt_gnu_mutable